requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
